+++ apps/blockchain-postgres-sync/src/lib/consumer/models/mod.rs
@@ -3,4 +3,4 @@ pub mod assets;
 pub mod block_microblock;
 pub mod candles;
 pub mod txs;
-pub mod waves_data;
+pub mod dcc_data;